requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# ESP/get_packets.py: 2
# ESP/processing.py: 15
Pillow == 7.2.0

# ESP/processing.py: 10
matplotlib == 3.2.2

# ESP/get_packets.py: 8
# ESP/mc.py: 8,9
# ESP/processing.py: 17
mcpi == 1.2.0

# ESP/processing.py: 14
mss == 6.0.0

# ESP/converter.py: 9
# ESP/get_packets.py: 5
# ESP/mc.py: 7
# ESP/processing.py: 9
numpy == 1.18.5

# ESP/converter.py: 8
# ESP/get_packets.py: 1
# ESP/processing.py: 8
opencv_python == 4.4.0.44

# ESP/processing.py: 16
pyscreenshot == 2.2

# ESP/config.py: 7
# ESP/get_packets.py: 3
requests == 2.24.0

# ESP/converter.py: 10
# ESP/mc.py: 10
scikit_image == 0.16.2
